Add files via upload
Diabetes prediction is a critical task for early intervention and management of the disease. This study focuses on the application of classification models for predicting diabetes. A comprehensive dataset comprising various clinical and demographic features related to diabetic patients is utilized. Preprocessing techniques are applied to prepare the data, followed by the implementation of several classification algorithms for better understandings. Performance evaluation is conducted using metrics such as accuracy, precision, recall, and F1-score. The results highlight the effectiveness of classification models in accurately predicting diabetes, facilitating timely intervention and personalized treatment plans. The findings of this study hold significant implications for healthcare professionals, enabling them to make informed decisions for diabetes prevention and management, thereby improving patient outcomes and quality of life